src/requirements.txt
transformers
protobuf
sentencepiece
sentence-transformers
python-dotenv
torch
huggingface_hub
bitsandbytes
accelerate
faiss-cpu